requirements.txt
httpx[http2]>=0.27.0
playwright>=1.40.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
loguru>=0.7.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
